More consistent interface / struct naming
Interfaces that describe structs (which mostly exist for testing, so
that a concrete struct can be substituted for a mock) should be
prefixed with an I (e.g. FrameLayer -> IFrameLayer).

Types that were named oddly to avoid conflicts with previous interface
names are now named what those interfaces were previously.