Day brightener
Dear future self, I'm leaving this here to brighten your day.
Consider this an apology for doing this in the first place, because
I have no doubt that you're reading this commit as part of a blame
because you're trying to debug some horrible, horrible bug. Just
know that I was thinking of you, and while I'm sorry, I'm not that
sorry. Instead of judging me, please think of all the horrible things
you've done to our further-in-the-future self, and then you will
understand why I've done what I've done.

Warmest regards,

Your past self.