Some code generation stuff
I was thinking about using this for command class parsing, but I
think that all of the different types (BYTE_STRUCT in particular)
are going to end up making it more complicated than it's really
worth, and we really don't need to support *so* many command classes
that it's really worth the time it's going to take to figure out how
to format everything.

I committed the XML definitions just in case (not a terrible reference
if you want to verify at the mask/shift values for a particular CC).