Skip generation of certain command classes
ZWAVE_CMD_CLASS is not a command class

Command classes containing a MARKER are too annoying to write a parser
for right now. Note to self: it might be worth considering just writing
a totally different parser template for commands that contain marker
(since it involves some lookahead and stuff).